requirements.txt
aiosqlite==0.17.0
aiomysql==0.2.0
alembic==1.13.1
anyio==4.3.0
asgiref==3.5.2
attrs==22.2.0
bcrypt==3.2.2
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
cryptography==37.0.2
dependency-injector==4.41.0
dnspython==2.6.1
ecdsa==0.17.0
email_validator==2.2.0
execnet==1.9.0
fastapi==0.110.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.13.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.4
loguru==0.6.0
Mako==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
orjson==3.10.6
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-forked==1.4.0
pytest-mock==3.9.0
pytest-xdist==2.5.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rich==13.7.1
rsa==4.8
shellingham==1.5.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.30
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.15
starlette==0.36.3
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.12
uvicorn==0.20.0
watchfiles==0.21.0
websockets==12.0
zipp==3.8.0
